# refactor: optimize port map to switch statement