Initializing huge maps as global variables causes a lot of allocations while importing the package (even if the respective feature is never used).

Rewrite to use `switch` statements and regen `iana_ports.go`.

As observed with `GODEBUG=inittrace=1`:

before:

```
init github.com/gopacket/gopacket/layers @0.52 ms, 1.7 ms clock, 395496 bytes, 120 allocs
```

after:

```
init github.com/gopacket/gopacket/layers @1.0 ms, 1.4 ms clock, 27760 bytes, 86 allocs
```

Time can be ignored, but reduction in memory allocation is visible.